Add csv version, change date format
Changed update_mysql.py and config.py to ask for the first date and number of days' worth of data to retrieve, instead of asking for the number of days ago and retrieving all data from that many days ago until yesterday.

Also added pendo_csv.py, which functions the same as update_mysql.py but writes to a collection of csv files instead of writing to a MySQL database.

Added usage instructions to README.